Add nested conditionals w/one statement each
It's difficult to understand the control flow in this code;
different questions are asked throughout at different levels,
and it's hard to follow.

We need to re-arrange the conditionals, but this is hard when
they're different shapes. It's possible to change this structure:

if x
  if y
    #thing
  end
end

into:

if y
  if x
    #thing
  end
end

but this can't be done easily if there are else blocks, or multiple
statements inside a conditional.

This commit isolates single statements inside conditionals, so
that we can start to move them around.