Revert "Express quality logic for normal items in one place"
This reverts commit 85aea1f70644a9762071b38c0f0bd9d2be0a41c7.

We struggled with where to go next, looking at this 'sell in' section:

  if item.name != 'Sulfuras, Hand of Ragnaros'
    item.sell_in = item.sell_in - 1
  end

We found it hard to see where to go next to make this sensible.

We want to try isolating the 'sell in' part of the code, but we
can't do that with the original commit's change (since that change
is the first change where not every conditional contains a single
statement).